// src/app/srvapi/models/ProcessorOperationArgumentDesc.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProcessorOperationArgumentDesc
 */
export interface ProcessorOperationArgumentDesc {
    /**
     * 
     * @type {string}
     * @memberof ProcessorOperationArgumentDesc
     */
    name?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ProcessorOperationArgumentDesc
     */
    optional?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ProcessorOperationArgumentDesc
     */
    type?: ProcessorOperationArgumentDescTypeEnum;
}

export function ProcessorOperationArgumentDescFromJSON(json: any): ProcessorOperationArgumentDesc {
    return ProcessorOperationArgumentDescFromJSONTyped(json, false);
}

export function ProcessorOperationArgumentDescFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProcessorOperationArgumentDesc {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': !exists(json, 'name') ? undefined : json['name'],
        'optional': !exists(json, 'optional') ? undefined : json['optional'],
        'type': !exists(json, 'type') ? undefined : json['type'],
    };
}

export function ProcessorOperationArgumentDescToJSON(value?: ProcessorOperationArgumentDesc | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'optional': value.optional,
        'type': value.type,
    };
}

/**
* @export
* @enum {string}
*/
export enum ProcessorOperationArgumentDescTypeEnum {
    FLOAT = 'FLOAT',
    INT = 'INT',
    STRING = 'STRING',
    DATE = 'DATE',
    FLOATLIST = 'FLOAT_LIST',
    INTLIST = 'INT_LIST',
    STRINGLIST = 'STRING_LIST',
    DATELIST = 'DATE_LIST',
    VOID = 'VOID'
}